requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
asyncmy==0.2.10
attrs==25.3.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
fastjsonschema==2.21.1
fonttools==4.58.5
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mistune==3.1.3
msgpack==1.1.1
multidict==6.6.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-socketio==5.13.0
pywin32==310
pyzmq==27.0.0
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.15.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
webrtcvad-wheels==2.0.14
wsproto==1.2.0
yarg==0.1.9
yarl==1.20.1
